backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
selenium==4.15.2
webdriver-manager==4.0.1
pyspellchecker==0.7.2
language-tool-python==2.7.1
openpyxl==3.1.2
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
